Enabling the PWM for R_ST_SRV with no servo connected will cause bad supply for its potentiometer's buffer.
This will make the buffer clamp to the power rail and output 3.3V~5V, the behaviour is unpredictable.

Thankfully this doesn't occur when there's a servo connected and it seems like only R_ST_SRV has this problem.

+++ Core/Inc/servo.h
@@ -18,8 +18,8 @@
  typedef struct{
     TIM_HandleTypeDef* timer;     // timer handler
     uint32_t           channel;   // timer's pwm channel (check .ioc file)
-    uint16_t           v_pot;     // internal potentiometer voltage
-    uint16_t           i_sense;   // servo's current draw
+    uint32_t           v_pot;     // internal potentiometer voltage
+    uint32_t           i_sense;   // servo's current draw
  } Servo;
 
  void run_servo(Servo* servo, uint8_t dutycycle);
